refacotr: 추상메서드가 아닌 메서드를 상속받지 못하도록 변경
