+++ src/main/java/chess/model/piece/pawn/Pawn.java
@@ -30,7 +30,24 @@ public abstract class Pawn extends Piece {
 
     protected abstract boolean isRightDistance(final Distance distance);
 
-    protected boolean isSatisfySpecialCondition(final Distance distance, final Color targetColor) {
+    public abstract Piece update();
+
+    @Override
+    public final boolean isPawn() {
+        return true;
+    }
+
+    @Override
+    public final boolean isEmpty() {
+        return false;
+    }
+
+    @Override
+    public final boolean isKing() {
+        return false;
+    }
+
+    protected final boolean isSatisfySpecialCondition(final Distance distance, final Color targetColor) {
         final int rank = Math.abs(distance.rank());
 
         final Direction direction = distance.findDirection();
@@ -47,7 +64,7 @@ public abstract class Pawn extends Piece {
         return Direction.isUpOrDown(direction) && targetColor.isEmpty();
     }
 
-    protected boolean isRightAttack(final Distance distance, final Color targetColor) {
+    protected final boolean isRightAttack(final Distance distance, final Color targetColor) {
         return hasEnemy(targetColor) && isDiagonalAttack(distance);
     }
 
@@ -58,14 +75,5 @@ public abstract class Pawn extends Piece {
     private boolean isDiagonalAttack(final Distance distance) {
         return Direction.isDiagonal(distance.findDirection())
                 && Math.abs(distance.rank()) == MINIMUM_DISTANCE;
-    }
-
-    public Piece update() {
-        return this;
-    }
-
-    @Override
-    public boolean isPawn() {
-        return true;
     }
 }
